requirements.txt
os
shutil
tkinter
zipfile
rarfile
py7zr
tarfile
gzip
bz2
pycabinet
tabulate
moviepy
forex-python
docx
PyPDF2
Pillow
pyttsx3
qrcode
